requirements.txt
prometheus-client==0.16.0
boto3==1.28.38
pyyaml==6.0